fix: useModel context null error — rootContainer is outside model provider
rootContainer is the outermost React wrapper, rendered BEFORE the
dataflow/model provider. Components inside it cannot use useModel().

- ErrorBoundary: revert to pure class component with own online/offline
  listeners (no hooks). ProLayout ErrorBoundary prop requires
  ComponentClass anyway.
- OfflineBanner: revert to own useState+useEffect listeners (same reason
  — outside model provider in rootContainer).
- Delete unused src/models/network.ts (no consumers left).
- Remove ErrorBoundaryClass named export, keep default export only.
- Fix Alert `title` → `message` prop (banner mode uses message).
