+++ src/app.tsx
@@ -14,7 +14,6 @@ import {
   AvatarDropdown,
   DocLink,
   ErrorBoundary,
-  ErrorBoundaryClass,
   Footer,
   LangDropdown,
   OfflineBanner,
@@ -144,7 +143,7 @@ export const layout: RunTimeLayoutConfig = ({
       : [],
     // Replace ProLayout's default ErrorBoundary with our offline-aware version,
     // so chunk load errors show friendly messages instead of "Something went wrong."
-    ErrorBoundary: ErrorBoundaryClass,
+    ErrorBoundary,
     menuHeaderRender: undefined,
     // 自定义 403 页面
     // unAccessible: <div>unAccessible</div>,
