+++ src/components/OfflineBanner/index.tsx
@@ -1,8 +1,22 @@
-import { getIntl, useModel } from '@umijs/max';
+import { getIntl } from '@umijs/max';
 import { Alert } from 'antd';
+import { useEffect, useState } from 'react';
 
 const OfflineBanner: React.FC = () => {
-  const { isOnline } = useModel('network');
+  const [isOnline, setIsOnline] = useState(
+    typeof navigator !== 'undefined' ? navigator.onLine : true,
+  );
+
+  useEffect(() => {
+    const handleOnline = () => setIsOnline(true);
+    const handleOffline = () => setIsOnline(false);
+    window.addEventListener('online', handleOnline);
+    window.addEventListener('offline', handleOffline);
+    return () => {
+      window.removeEventListener('online', handleOnline);
+      window.removeEventListener('offline', handleOffline);
+    };
+  }, []);
 
   if (isOnline) return null;
